+++ WebI/trabalhofinal/src/useCases/user/deleteUser.ts
@@ -1,16 +1,28 @@
 import { PrismaClient, } from "@prisma/client";
+import { DeletedPresenter, IPresenter, NotFoundPresenter } from "../../presenters";
+import { GetUserUseCase } from "./getUser";
 
-const prisma = new PrismaClient(); 
+const prisma = new PrismaClient();
 
-export class DeleteUserUseCase{
-    constructor(){}
+export class DeleteUserUseCase {
+  constructor() { }
 
-    async handle(id: string){
+  async handle(id: string): Promise<IPresenter> {
 
-        await prisma.user.delete({
-            where: {
-              id: id,
-            },
-          });
+
+    const useCaseGet = new GetUserUseCase();
+    const userExists = await useCaseGet.handle(id);
+
+    if (!userExists) {
+      return new NotFoundPresenter({
+        message: 'User not found!',
+      });
     }
+    await prisma.user.delete({
+      where: {
+        id: id,
+      },
+    });
+    return new DeletedPresenter();
+  }
 }
